requirements.txt
Django==2.0.3
django-extensions==2.0.6
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2-binary==2.7.4
